requirements.txt
aiogram>=3.0.0,<4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
apscheduler>=3.10.0,<4.0.0
pytz>=2024.1
aiofiles>=23.1.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
